[CI] move rayservice_*.go into e2erayservice folder
