+++ ray-operator/test/e2erayservice/testdata/rayservice.autoscaling.yaml
