+++ ray-operator/test/e2erayservice/testdata/rayservice.static.yaml
